feat: add login and registration pages, update routing
